requirements.txt
# Django==4.1.1
# psycopg2-binary==2.9.3
# python-dotenv
langchain
openai
streamlit
PyPDF2
# Flask
# wikipedia
# chromadb
tiktoken
faiss-cpu
python-dotenv
